+++ spec/request_validation/parameter_validation_spec.rb
@@ -31,7 +31,7 @@ RSpec.describe 'Parameter validation' do
   describe '#call' do
     let(:params) do
       {
-        'term' => 'Oscar'
+        term: 'Oscar'
       }
     end
 
@@ -57,7 +57,7 @@ RSpec.describe 'Parameter validation' do
     end
 
     it 'returns 400 if query parameter is missing' do
-      params.delete('term')
+      params.delete(:term)
       get path, params
 
       expect(last_response.status).to eq 400
@@ -124,13 +124,13 @@ RSpec.describe 'Parameter validation' do
 
       let(:params) do
         {
-          'term' => 'Oscar',
-          'filter' => { 'tag' => 'dogs', 'id' => '1', 'other' => 'things' }
+          term: 'Oscar',
+          filter: { tag: 'dogs', id: '1', other: 'things' }
         }
       end
 
       it 'returns 400 if nested[parameter] is missing' do
-        params['filter'].delete('tag')
+        params[:filter].delete(:tag)
         get path, params
 
         expect(last_response.status).to eq 400
@@ -150,7 +150,7 @@ RSpec.describe 'Parameter validation' do
       let(:raise_error_option) { true }
 
       it 'raises an error if query parameter is missing' do
-        params.delete('term')
+        params.delete(:term)
         message = 'Query parameter invalid: is missing required properties: term'
         expect do
           get path, params
@@ -175,7 +175,7 @@ RSpec.describe 'Parameter validation' do
         get path, params.merge(limit: '100')
 
         expect(last_response.status).to eq(200), last_response.body
-        expect(last_params['limit']).to eq 100
+        expect(last_params[:limit]).to eq 100
 
         get path, params.merge(limit: 'invalid')
         expect(last_response.status).to eq(400)
@@ -188,7 +188,7 @@ RSpec.describe 'Parameter validation' do
         get path, params.merge(weight: '1.5')
 
         expect(last_response.status).to eq(200), last_response.body
-        expect(last_params['weight']).to eq 1.5
+        expect(last_params[:weight]).to eq 1.5
 
         get path, params.merge(limit: 'invalid')
         expect(last_response.status).to eq(400)
@@ -200,22 +200,22 @@ RSpec.describe 'Parameter validation' do
       it 'converts to boolean' do
         get path, params.merge(starred: 'true')
         expect(last_response.status).to eq(200), last_response.body
-        expect(last_params['starred']).to eq true
+        expect(last_params[:starred]).to eq true
 
         get path, params.merge(starred: 'false')
         expect(last_response.status).to eq(200), last_response.body
-        expect(last_params['starred']).to eq false
+        expect(last_params[:starred]).to eq false
 
         get path, params.merge(starred: 'wrong')
         expect(last_response.status).to eq(400)
-        expect(last_params['starred']).to eq 'wrong'
+        expect(last_params[:starred]).to eq 'wrong'
       end
 
       it 'converts nested params' do
         get path, params.merge(filter: { id: '100', tag: 'foo' })
 
         expect(last_response.status).to eq(200), last_response.body
-        expect(last_params['filter']['id']).to eq 100
+        expect(last_params[:filter][:id]).to eq 100
       end
     end
   end